+++ vkconfig/CHANGELOG.md
@@ -16,6 +16,9 @@
 ### Features:
 - Add per layers-configuration user-defined layers paths #1711
 
+### Improvements:
+- Add reordering of configurations when renaming a configuration
+
 ## [Vulkan Configurator 2.4.6](https://github.com/LunarG/VulkanTools/tree/sdk-1.3.216.0) - June 2022
 
 ### Features: